First Version!
Already a really fun project! I haven't made a functioning Chrome
extension before this project, so I'm really happy that I made this in just a day, with a little more research than before, and
that it already works great.

I submitted a bug report suggesting to add the <meta name="theme-color">
tag to the web app's <head>, but their team somehow had no idea that
it exists, kind of confusing but ok. So I decided to fix it myself!